requirements/dev.txt
## Development requirements

-r common.txt
-r docs.txt

virtualenv
#virtualenvwrapper


## Testing
pytest
pytest-django
pytest-xdist
pytest-cov
pytest-capturelog
hypothesis
tox
coverage
pdbpp


# Fake data creation
faker
imagefactory


## Browser Testing
html5lib
selenium
